Made invoice downloadable
