+++ styles/scss/_invoice.scss
@@ -7,16 +7,18 @@
   background-color: white;
   color: #111;
   width: 100%;
-  height: 100%;
-  padding: 0.8rem;
+  // height: 100%;
+  padding: 0.6rem 0.4rem;
   ul {
     list-style: none;
   }
 }
 
 .invoice.view {
-  border: 1px solid black;
-  padding: 1rem;
+  width: 80%;
+  margin: 0 auto;
+  padding: 0.8rem;
+  font-size: 1.55rem;
 }
 
 .invoice-header {
@@ -40,10 +42,16 @@
 table {
   margin: 7.5vh 0;
   width: 100%;
-  text-align: left;
-
   tr {
-    height: 2rem;
+    text-align: left;
+    vertical-align: top;
+    margin: auto;
+    height: 3rem;
+  }
+
+  tr th {
+    text-align: left;
+    vertical-align: top;
   }
 
   .hide-on-show {
